docs: remove use of &nbsp; in docs
Some docs pages were using <p>&nbsp;</p> to add arbitrary whitespace
in the page. This is something that should be done by CSS if needed,
but it is not needed here, so delete it.

There was also use of <td>&nbsp;</td> which adds no value at all
when we have CSS to prettify tables.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>