Revert "travis: Limit git depth to 5 commits"
Turns out a build job can be stuck waiting for a macOS worker to
become available for a pretty long time: if more than 5 commits
have been pushed in the meantime, the clone will be too shallow
for the worker to find the commit it's supposed to verify, and
the build job will fail.

See https://travis-ci.org/libvirt/libvirt/jobs/277244110 for an
example of the failure described.

This reverts commit 2e975abdc9bbc9e965486e8486cc17a771cdaeb3.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>
Reviewed-by: Daniel P. Berrange <berrange@redhat.com>